requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
apscheduler==3.10.4
aiosqlite==0.19.0
openpyxl==3.1.2